+++ backend/src/rag_utils/hanhwa-testdata.pdf.placeholder
